requirements.txt
groq
moviepy
transformers
datasets
huggingface_hub
torch
opencv
tensorflow
accelerate
tf-keras
pillow